loginAndSingnup
Pages Login and Register upgraded to CSS variables. 'navigate' function added to Buttons, user validation with 'yup' added (warnings with 'react-toastify'), redirection added (temporarily to '/home') if the user is logged in. Inputs are saved in the store (using reducers), loading when rendering and reset to the initial state when the form is submitted. OnSubmit, all credentials are sent to operations (session). Additionally, the '--color-blur' variable was added to the CSS variables file. Two reducers (resetLoginForm and resetRegisterForm) were added to sessionSlice.js, necessary to reset the store of the forms.
Additionally, the files App.jsx and main.jsx were modified to help test the functionalities of the pages. They will be restored to their original values before the Pull Request.